+++ server/component/style/textarea/TextareaView.php
@@ -52,6 +52,9 @@ class TextareaView extends FormFieldView
         }
         if($this->value === null)
             $this->value = $this->default_value;
+        if($this->locked_after_submit == 1){
+            $this->locked_after_submit = $this->value ? 1 : 0;
+        }
         $css = ($this->label == "") ? $this->css : "";
         $required = ($this->is_required) ? "required" : "";
         require __DIR__ . "/tpl_textarea.php";