+++ CHANGELOG.md
@@ -10,6 +10,7 @@
  - add `searchBuilder` for `scheduledJobs`;
  - add `searchBuilder` for `users`;
  - add header position for page edit in the old UI. Fix page editing in old UI to remove pages from menu
+ - saving form with `ajax` enabled will refresh the inputs based on the values 
 
 # v5.1.5
 ### Bugfix